Expose API to control properties over time
This initial version only supports controlling properties of mixer
slots, the API is generic however and should be easy to implement
for other nodes (eg Filter nodes when implemented).

Any property exposed by the mixer pads can be controlled through
that API, provided it is writable, readable if interpolating,
numerically-typed, or boolean / string if mode is "Set".

+ Update MixerInfo to include control points [API break]

+ Add sample script to demonstrate usage, `crossfade.py`